+++ libs/qt/source/ftrack_qt/widgets/frames/asset_version_selection.py
@@ -4,7 +4,7 @@
 from Qt import QtWidgets, QtCore, QtGui
 
 from ftrack_qt.widgets.selectors.version_selector import VersionSelector
-from ftrack_qt.widgets.thumbnails import OpenAssetVersionThumbnail
+from ftrack_qt.widgets.thumbnails import AssetVersionThumbnail
 
 
 class AssetVersionSelection(QtWidgets.QFrame):
@@ -51,7 +51,7 @@ class AssetVersionSelection(QtWidgets.QFrame):
         self.layout().setSpacing(5)
 
     def build(self):
-        self._thumbnail_widget = OpenAssetVersionThumbnail()
+        self._thumbnail_widget = AssetVersionThumbnail()
         self._thumbnail_widget.setScaledContents(True)
         self._thumbnail_widget.setMinimumSize(57, 31)
         self._thumbnail_widget.setMaximumSize(57, 31)
