+++ libs/qt/source/ftrack_qt/widgets/selectors/__init__.py
@@ -1,9 +1,8 @@
 from ftrack_qt.widgets.selectors.list_selector import ListSelector
 from ftrack_qt.widgets.selectors.context_selector import ContextSelector
-from ftrack_qt.widgets.selectors.asset_selector import AssetSelector
-from ftrack_qt.widgets.selectors.status_selector import StatusSelector
-from ftrack_qt.widgets.selectors.version_selector import VersionSelector
-from ftrack_qt.widgets.selectors.open_asset_selector import OpenAssetSelector
-from ftrack_qt.widgets.selectors.open_asset_selector import (
+from ftrack_qt.widgets.selectors.asset_selector import (
+    OpenAssetSelector,
     PublishAssetSelector,
 )
+from ftrack_qt.widgets.selectors.status_selector import StatusSelector
+from ftrack_qt.widgets.selectors.version_selector import VersionSelector
