--- libs/qt/source/ftrack_qt/widgets/selectors/open_asset_selector.py
@@ -1,197 +1,0 @@
-# :coding: utf-8
-# :copyright: Copyright (c) 2014-2023 ftrack
-
-import logging
-
-from Qt import QtWidgets, QtCore, QtGui
-
-from ftrack_qt.widgets.frames import (
-    AssetVersionCreation,
-    AssetVersionSelection,
-    NewAssetInput,
-)
-from ftrack_qt.widgets.lists import AssetList
-
-
-class OpenAssetSelector(QtWidgets.QWidget):
-    '''Widget for choosing an existing asset and asset version, or input asset
-    name for creating a new asset, depending on mode.'''
-
-    assets_added = QtCore.Signal(object)
-    '''Signal emitted when assets are added, with list of asset entities as argument'''
-
-    version_changed = QtCore.Signal(object)
-    '''Signal emitted when version is changed, with assetversion entity as
-    argument (version select mode)'''
-
-    selected_item_changed = QtCore.Signal(object)
-    '''Signal emitted when an item is selected, with asset_version id
-    argument'''
-
-    def __init__(
-        self,
-        parent=None,
-    ):
-        '''
-        Initialise asset selector widget.
-
-        :param mode: The mode of operation.
-        :param fetch_assets: Callback to fetch assets
-        :param session: ftrack session, required for thumbnail load.
-        :param fetch_assetversions: Callback to fetch asset version for a specific
-        asset.
-        :param parent:
-        '''
-        super(OpenAssetSelector, self).__init__(parent=parent)
-        self.logger = logging.getLogger(
-            __name__ + '.' + self.__class__.__name__
-        )
-
-        self._list_and_input = None
-        self._asset_list = None
-        self._new_asset_input = None
-
-        self.selected_index = None
-
-        self.pre_build()
-        self.build()
-        self.post_build()
-
-    def pre_build(self):
-        main_layout = QtWidgets.QVBoxLayout()
-        main_layout.setContentsMargins(0, 0, 0, 0)
-        self.setLayout(main_layout)
-
-    def build(self):
-        self._asset_list = AssetList(AssetVersionSelection)
-
-        self.layout().addWidget(self._asset_list)
-
-    def post_build(self):
-        self._asset_list.assets_added.connect(self._on_assets_added)
-        self._asset_list.version_changed.connect(self._on_version_changed)
-        self._asset_list.selected_item_changed.connect(
-            self._on_selected_item_changed
-        )
-
-    def _on_assets_added(self, assets):
-        self.assets_added.emit(assets)
-
-    def set_assets(self, assets):
-        if not assets:
-            self._asset_list.hide()
-        else:
-            self._asset_list.show()
-        self._asset_list.set_assets(assets)
-
-    def _on_version_changed(self, version):
-        self.version_changed.emit(version)
-
-    def _on_selected_item_changed(self, index, version):
-        self.selected_index = index
-        self.selected_item_changed.emit(version)
-
-
-class PublishAssetSelector(OpenAssetSelector):
-    '''Widget for choosing an existing asset and asset version, or input asset
-    name for creating a new asset, depending on mode.'''
-
-    VALID_ASSET_NAME = QtCore.QRegExp('[A-Za-z0-9_]+')
-
-    new_asset = QtCore.Signal(object)
-
-    def __init__(
-        self,
-        parent=None,
-    ):
-        '''
-        Initialise asset selector widget.
-
-        :param mode: The mode of operation.
-        :param fetch_assets: Callback to fetch assets
-        :param session: ftrack session, required for thumbnail load.
-        :param fetch_assetversions: Callback to fetch asset version for a specific
-        asset.
-        :param parent:
-        '''
-        self.validator = QtGui.QRegExpValidator(self.VALID_ASSET_NAME)
-        self.placeholder_name = "Asset Name..."
-
-        super(PublishAssetSelector, self).__init__(parent=parent)
-
-    def build(self):
-        self._list_and_input = AssetListAndInput()
-
-        self._asset_list = AssetList(AssetVersionCreation)
-
-        self._list_and_input.add_asset_list(self._asset_list)
-
-        # Create new asset
-        self._new_asset_input = NewAssetInput(
-            self.validator, self.placeholder_name
-        )
-        self._list_and_input.layout().addWidget(self._new_asset_input)
-
-        self.layout().addWidget(self._list_and_input)
-
-    def post_build(self):
-        super(PublishAssetSelector, self).post_build()
-        self._new_asset_input.text_changed.connect(self._on_new_asset)
-
-    def _on_new_asset(self, asset_name):
-        '''New asset name text changed'''
-        self.selected_index = None
-        self.selected_item_changed.emit(None)
-        is_valid_name = self.validate_name(asset_name)
-        if is_valid_name:
-            self.new_asset.emit(asset_name)
-        else:
-            self.new_asset.emit(None)
-
-    def validate_name(self, asset_name):
-        '''Return True if *asset_name* is valid, also reflect this on input style'''
-        is_valid_bool = True
-        # Already an asset by that name
-        if self._asset_list.assets:
-            for asset_id, asset_dict in self._asset_list.assets.items():
-                if asset_dict['name'].lower() == asset_name.lower():
-                    is_valid_bool = False
-                    break
-        if is_valid_bool and self.validator:
-            is_valid = self.validator.validate(asset_name, 0)
-            if is_valid[0] != QtGui.QValidator.Acceptable:
-                is_valid_bool = False
-            else:
-                is_valid_bool = True
-        self._new_asset_input.set_valid(is_valid_bool)
-        return is_valid_bool
-
-    def set_default_new_asset_name(self, name):
-        self._new_asset_input.set_default_name(name)
-
-
-class AssetListAndInput(QtWidgets.QWidget):
-    '''Compound widget containing asset list and new asset input'''
-
-    def __init__(self, parent=None):
-        super(AssetListAndInput, self).__init__(parent=parent)
-        self.setLayout(QtWidgets.QVBoxLayout())
-        self.layout().setContentsMargins(0, 0, 0, 0)
-        self.layout().setSpacing(0)
-
-    def add_asset_list(self, asset_list):
-        '''Add *asset_list* widget to widget'''
-        self._asset_list = asset_list
-        self.layout().addWidget(asset_list)
-
-    def resizeEvent(self, event):
-        '''(Override)'''
-        self.size_changed()
-
-    def size_changed(self):
-        '''Resize asset list to fit widget, to prevent unnecessary scrolling'''
-        self._asset_list.setFixedSize(
-            self.size().width() - 1,
-            self._asset_list.sizeHintForRow(0) * self._asset_list.count()
-            + 2 * self._asset_list.frameWidth(),
-        )
