+++ libs/qt/source/ftrack_qt/widgets/thumbnails/__init__.py
@@ -3,6 +3,4 @@ from ftrack_qt.widgets.thumbnails.base import ThumbnailBase
 from ftrack_qt.widgets.thumbnails.context import ContextThumbnail
 from ftrack_qt.widgets.thumbnails.ellipse import EllipseThumbnailBase
 from ftrack_qt.widgets.thumbnails.user import UserThumbnail
-from ftrack_qt.widgets.thumbnails.open_asset_version import (
-    OpenAssetVersionThumbnail,
-)
+from ftrack_qt.widgets.thumbnails.session_base import SessionThumbnailBase
