+++ libs/qt/source/ftrack_qt/widgets/thumbnails/context.py
@@ -6,10 +6,10 @@ import urllib.request, urllib.parse, urllib.error
 
 from Qt import QtCore, QtGui, QtWidgets
 
-from ftrack_qt.widgets.thumbnails.base import ThumbnailBase
+from ftrack_qt.widgets.thumbnails.session_base import SessionThumbnailBase
 
 
-class ContextThumbnail(ThumbnailBase):
+class ContextThumbnail(SessionThumbnailBase):
     '''Context thumbnail widget'''
 
     def _download(self, reference):
